feat: Add support for new home delivery and print in store features
The plugin now supports the new home delivery feature, returning the shipment label in the shipment data. To print the label at home, add "type"="home" in the shipping option metadata created within the mondialrelay provider.

Additionally, the plugin now supports the new print in store feature, returning the pickup point reference in the shipment data. To print the label in the store, add "print"="in_store" in the shipping option metadata created within the mondialrelay provider.

If the type is not set to "home" or the print is not set to "in_store", the default delivery mode or A4 PDF label will be used, respectively.